refactor: Consolidate duplicate Gemini API logic into shared utility
Extracted common code from narrative/generate and narrative/choices routes
into a new processGeminiTextRequest function in apiHelpers.ts.

Changes:
- Created processGeminiTextRequest() to handle the complete flow:
  - Rate limiting
  - Request validation
  - API key validation
  - Gemini API call with configurable options
  - Response parsing and error handling
- Simplified both route files from ~130 lines to 12 lines each
- Eliminated ~230 lines of duplicate code
- Added GeminiTextResponse and GeminiTextOptions interfaces

The behavior remains identical, just consolidated for maintainability.
